Update jedis to 3.6.0

+++ module/infra/src/main/scala/kindleclock/infra/cache/redis/RedisCacheClientJedisImpl.scala
@@ -34,7 +34,7 @@ class RedisCacheClientJedisImpl @Inject() (
     Future(
       binaryJedis.setex(
         keyName.getBytes(charset),
-        expiration.toSeconds.toInt,
+        expiration.toSeconds,
         a.toByteArray
       ) == "OK"
     )
